move sql to top
